+++ public/main/inc/ajax/agenda.ajax.php
@@ -35,6 +35,9 @@ switch ($action) {
         if (!$agenda->getIsAllowedToEdit()) {
             break;
         }
+        if (false === Security::check_token('get')) {
+            exit;
+        }
         $add_as_announcement = $_REQUEST['add_as_annonuncement'] ?? null;
         $title = $_REQUEST['title'] ?? null;
         $content = $_REQUEST['content'] ?? null;
@@ -59,8 +62,13 @@ switch ($action) {
         if (!$agenda->getIsAllowedToEdit()) {
             break;
         }
+        if (false === Security::check_token('get')) {
+            exit;
+        }
         $id_list = explode('_', $_REQUEST['id']);
         $id = $id_list[1];
+        $title = $_REQUEST['title'] ?? null;
+        $content = $_REQUEST['content'] ?? null;
         $agenda->editEvent(
             $id,
             $_REQUEST['start'],
@@ -74,6 +82,9 @@ switch ($action) {
         if (!$agenda->getIsAllowedToEdit()) {
             break;
         }
+        if (false === Security::check_token('get')) {
+            exit;
+        }
         $id_list = explode('_', $_REQUEST['id']);
         $id = $id_list[1];
         $deleteAllEventsFromSerie = isset($_REQUEST['delete_all_events']) ? true : false;
@@ -83,6 +94,9 @@ switch ($action) {
         if (!$agenda->getIsAllowedToEdit()) {
             break;
         }
+        if (false === Security::check_token('get')) {
+            exit;
+        }
         $minute_delta = $_REQUEST['minute_delta'];
         $id = explode('_', $_REQUEST['id']);
         $id = $id[1];
@@ -91,6 +105,9 @@ switch ($action) {
     case 'move_event':
         if (!$agenda->getIsAllowedToEdit()) {
             break;
+        }
+        if (false === Security::check_token('get')) {
+            exit;
         }
         $minute_delta = $_REQUEST['minute_delta'];
         $allDay = $_REQUEST['all_day'];
@@ -124,7 +141,7 @@ switch ($action) {
         echo $events;
         break;
     case 'get_user_agenda':
-        //Used in the admin user list
+        // Used in the admin user list.
         api_protect_admin_script();
 
         if (api_is_allowed_to_edit(null, true)) {
